CPU Copy had too many instantiations - Splitting into 2 files
* copy.cpp and padarray.cpp
* The copy.hpp file remains the same
